LayoutTests/fast/xmlhttprequest/xmlhttprequest-responseXML-xml-text-responsetype-expected.txt

This tests the XMLHttpRequest responseXML loading an XML document with responseType "text".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.responseXML threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
